+++ src/repositories/teacherQualification/index.ts
@@ -1,0 +1,1 @@
+export * from "@/repositories/teacherQualification/teacherQualification.repository";