+++ src/services/teacher/dto/index.ts
@@ -1,0 +1,3 @@
+export * from "@/services/teacher/dto/teacher.dto";
+export * from "@/services/teacher/dto/teacherQualification.dto";
+export * from "@/services/teacher/dto/teacherAssignment.dto";