+++ src/services/teacher/index.ts
@@ -1,0 +1,2 @@
+export * from "@/services/teacher/teacher.service";
+export * from "@/services/teacher/dto";